+++ data/lua/utils/reload.lua
@@ -1,11 +1,11 @@
 local reload = {}
 
-function reload:Reload(module_name)  
+function reload.Reload(module_name)  
     package.loaded[module_name] = nil  
     require(module_name)  
 end
 
-function reload:ReloadAll()
+function reload.ReloadAll()
     local loaded_modules = {}
     for module_name, _ in pairs(package.loaded) do
         table.insert(loaded_modules, module_name)
